+++ src/libs/ascent/hola/ascent_hola.cpp
@@ -25,6 +25,7 @@
 // ascent includes
 //-----------------------------------------------------------------------------
 #include <ascent_logging.hpp>
+#include <ascent_logging_old.hpp>
 
 #include <fstream>