+++ src/libs/ascent/runtimes/flow_filters/ascent_runtime_dray_filters.cpp
@@ -451,6 +451,7 @@ parse_color_table(const conduit::Node &color_table_node)
         const Node &peg = itr.next();
         if(!peg.has_child("position"))
         {
+            // FIXME: This should be an error
             ASCENT_WARN("Color map control point must have a position");
         }